tomcat/src/main/java/org/apache/catalina/SessionManager.java
package org.apache.catalina;

import java.util.Map;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;

import static java.util.Objects.requireNonNull;
import static java.util.Optional.ofNullable;

public class SessionManager {

    private SessionManager() {
    }

    private static final Map<String, Session> SESSIONS = new ConcurrentHashMap<>();

    public static void add(final Session session) {
        SESSIONS.put(session.getId(), session);
    }

    public static Optional<Session> findSession(final String id) {
        requireNonNull(id, "세션 아이디가 null 입니다.");
        return ofNullable(SESSIONS.get(id));
    }

}
